requirements.txt
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
tiktoken>=0.5.0
rouge-score>=0.1.2
bert-score>=0.3.13
rich>=13.0.0
pydantic>=2.0.0
numpy>=1.24.0
jsonlines>=3.1.0
pytest>=7.0.0
matplotlib>=3.0.0
python-pptx>=0.6.0
